Revert "Created ConstructorBasedMongoPersistenceModule which is intended for the Non-DI usage"
