--- src/main/java/org/atlasapi/media/channel/ServiceChannelStore.java
@@ -1,9 +1,0 @@
-package org.atlasapi.media.channel;
-
-public interface ServiceChannelStore extends ChannelStore {
-
-    void start();
-
-    void shutdown();
-
-}
